Add test for problem with header file dependency of device adapters
There currently is a problem with the device adapters where if you
define the device adapter (particularly the instance of the algorithms)
after you define a templated function that uses it, you will get a
compile error.

Because the actual problem is not fixed, this results in a compile
error.  When the dependency issue is fixed, this test should pass.