Copy non-const portals to execution for in place operations
The ArrayHandle::PrepareForInPlace was just calling its own
PrepareForInput to pass data to the execution environment.  When using
the shared memory array manager, this did not actually prepare the
execution array for writing.

This change makes a special in place operation in ArrayManagerExecution
to make sure the data is copied to a writable array.  I'm surprised we
have not encountered this problem before.