Add handling of TPL licenses.
Added a note at the bottom of LICENSE.txt that some files are third
party and have independent licenses and lists them.  The check copyright
test now also grabs this list from the LICENSE.txt file and uses it to
skip over the checking of the copyright in those files.