Initialize members of exec cells to suppress warnings
When all warnings are on, there was a warning created by the copy
constructors of the execution cell types.  This was because the default
constructor of these classes did not set all the values in internal
tuples.  Changed the default constructor to initalize the tuples to get
around this error.

Also added some EXPORT macros to make things more explicit on where
functions are called.