Fix VectorTraits::HasMultipleComponents for tuples of size 1.
Previously, VectorTraits assumed that any class that defined
NUM_COMPONENTS and the other methods of dax::Tuple had multiple
components (i.e. HasMultipleComponents was typedefed to
VectorTraitsTagMultipleComponents).  However, it is possible to have
tuples of size 1, in which case this tag should really be
VectorTraitsTagSingleComponent.  In fact, there was even a comment
stating the possiblity and the potential need to change.

Well, turns out that tuples of size 1 are created in Dax (specifically
for index lists of vertex cells), and that the vector operations were
misbehaving because they assumed that they had at least 2 entries
because of the tag.

This change fixes the problem by using a suplemental template
specialized on the number of components.