GenerateTopology supports VisitIndex exec argument.
We convert the VisitIndex exec argument to a field which
holds the value that represents the number of times we have
already visited the input cell if we are running in serial.

You can think of the VisitIndex as representing the variable j
when you are doing a for loop i over all input cells and an
inner for loop j over all output cells that i cell will generate.