Implementing ArrayPortalPermutation with regular array for Key and Value
- This is a first pass on this array portals and basically uses array.
- Also convenience we are using Implicit arrays (this needs to change).
- ArrayPortalPermutation needs to take in two ArrayPortals instead of arrays.
- Unit test is added as well.

WIP: Trying to make permutation with implicit array

Simplifying ArrayContainerControlPermutation.

This is based on the implicit array. The container comprises of two arrays one
with the Key and other with the Value. The Key holds the index into Value array
and the corresponding value. So for example if the Key array = [0,2,3,5] Value
array = [8,6,4,9,8,3]. The array[2] = Value[Key[2]] = [3] = 9.

NOTE: This is only a first working implementation and more fixes may be needed

Extending test with different sizez for Key and Value arrays.

Key/Value arrays were the same in the previous test. This test is a little
better and test if the Permutation work for arrays of different sizes.

Changing API and adding assers for ArrayContainerControlPermutation

- We now need to specify both the key and value data array sizes.
- The tests are chaned accordingly
- A dax assert checks if the index in the key is relevant at construction time

WIP: Testing another permutation container option