Update VTK writer for Threshold benchmark.
The code to write the result of the threshold benchmark to a VTK file
was way out of date.  The only reason no one noticed is because its
instantiation was commented out and therefore the compile errors were
hidden.  However, it is useful to occasionally uncomment the writing to
verify that the output is correct.