Make triTable a const instead of a static.
The static declaration was giving a warning about an unused variable.
It may also not be a great idea to put a static table in a header anyway
as that can cause the compiler and linker problems trying to resolve
which one to use.  Plus, you don't want to change the data, and using
const might give the compiler better optimizations.