Introduce ability to choose scheduler based on worklet type.
DetermineScheduler determines at compile time what scheduler tag
is best option for each worklet. For each scheduler type we
implement a class that determines if it can schedule a given
worklet type. To determine which of scheduler is choosen, we
add all these choices to a mpl vector and extract the first
scheduler that returns it can execute the worklet type.

This solution was choosen because it is really easy to extend
the vector to  support more worklets, unlike using mpl::if_
which is incredibly hard to read when you have multiple levels of
nested if's.

This form of vector and ranking could be used to implement a
multi step scheduler, or a system where scheduler rank how well
they support a worklets requested arguments, etc

The other thing important part of this design is that the default
scheduler always states that it can schedule every worklet, so
it must always be the last scheduler in the vector as it
is the fail safe and if it is infront of specialized schedulers
we lose the entire purpose of this class.