Implement the default scheduling algorithm based on tag dispatch.
We have refactored the default implementation to now work
in the tag dispatching system. Instead of the previous two
header file implementation I have moved to a single file.
Also new is better user argument checking and moving
the CollectCount and CreateExecutionResource classes
to different header files, as they are shared classes.