AddVisitIndexArg generates visitindex field and new cont sig.
The VisItIndex exec signature requires the control signature
of the worklet to be extended to include an extra field. The
reason for this is that only items that relate to a control
signature argument are uploaded to the exec env, and accessable
in the exec env.

The visitIndex is an exec signature only placeholder, so therefore
to get it too work properly we have to modify the control signature
of the users worklet. Actually modifying the signature is impossible,
so we inherit from the users worklet and set the derived worklets
signature to have the extended signature.

To simplfy the code in ScheduleGenerateTopology, every generate
worklet is extended with VisitIndexArg. The trick is that when
the user doesn't explictly request the argument we only upload
a single constant integer value instead of a real array so we
impart zero overhead when we don't have a request for the visitIndexArg.