VerifyUserArgLengths verifies worklets are called properly.
VerifyUserArgLengths verifies at compile time that worklets
are called with the exact number of arguments they requested in the
control signature. This is a unique class so that each scheduling
algorithm can reuse the code.

We remove the original implementation from Bindings.h as
it wasn't as clear to users when reading the assert message
why it was happening. Now with it in the scheduler and
with the added comments it should be very easy to understand
why they are getting the failure.