Reverse order of CellField template arguments.
Put the FieldType first and CellTag second.  This is more consistent
with other classes in Dax that contain data where the first argument has
the data type and subsequent arguments alter how they are stored.  For
example, the Tuple argument has the type first and size second.  Even
though classes like UnstructuredGrid have CellTag, CellTag is really
telling the type stored in the grid (whereas in FieldType it alters how
it is stored).