Make Vector# and Id3 typedefs instead of subclasses.
Rather than have separate Vector and Id3 classes that subclass Tuple,
make specializations of Tuple for these counts and typedef these
symbols.  This gets around potential problems with converting between a
Vector and Tuple, which are really the same thing.

One downside of this change is that error messages will now always
report Tuple<float,3> instead Vector3, for example.  Another is that
template resolution is a bit more fragile.  The compiler had problems
determining the correct version of the == operator when using some funky
iterators.