Make all VectorTraits specialized.
Rather than have a generic VectorTraits that assumes that all
unspecialized classes have the same accessors as Tuple, make an
undefined prototype and a partial specialization for Tuple types.

One big advantage of this approach is that if VectorTraits is used on
something that is not really a vector (which is common when template
parameters get messed up), you get an informative warning that
VectorTraits<class> is not defined, rather than some odd warning about a
random class missing the [] operator or something like that.