Refactor TopologyCell to properly work with FieldMap.
Instead of having TopologyCell return by default the tag of the cell type,
we instead use a custom binding called BindCellTag to have the desired behavior
that _N as a topo execution signature returns the tag. When a user uses
Topo(_N) we do a direct binding to TopologyCell and get the cell vertices.

In the future I want to look at redoing the entire functor argument/binding
creation to see if we can get even more efficient cell vertice creation
and sharing.