Adding in new worklet type call InterpolatedCell.
InterpolatedCell worklet generates a cell where instead of having
pure vertices, we instead have the two vertex ids and the weight
between those two ids to create the new point.

This is a really important cell type to have so that we can generate
marching cubes with and without topology resolution. If we
have resolution enabled, the ids plus weight become the unique
key we can resolve duplicate based on.

This also allows us to store the interpolated information of
the marching cubes classification and easily apply it as the
coordinates change, but while the iso values stay constant.