Remove templates that marching cubes does not support.
The classify and topology worklets of marching cubes had templates that
allowed them to accept any kind of cell for inputs and outputs.  The
problem is that the code really only worked with hexahedra as input and
triangles as output.  This change removes the templates so that you get
a compile error if you try to use the wrong cell types.

Note that voxels are supported for input but hexahedra is not.  I have
not added it simply because I have no test for it, but it should
eventually be added and that should be straightforward.