Add cell traits to identify topology type and equivalent cells.
These new cell traits allow you to identify whether the cell comes from
a uniform or unstructured grid.  (New topology types are easily added.)
You can also identify a cononical cell type to make it easier to write,
for example, a function that operates on all hexahedra types (i.e.
CellTagVoxel and CellTagHexahedra and perhaps a curvilinear in the
future).

The marching cubes worklet was updated to support hexahedra in
unstructured grids.  The error messages that occur when the wrong type
is used are also more friendly.

The testing framework has also changed slightly.  The "try all"
functions have been consolidated for everything that has to do with a
cell.