Add a signature modifier for permuted domains.
This allows the domain checker to tell the difference between a regular
field and one that should be permuted.  It will also give the concept
mat the ability to change the concept when the domain is permuted (not
implemented yet).

Also expanded the UnitTestWorkletThreshold test to actually check the
topology of the output.  This is to make sure the changes I added to not
mess anything up.