Field Portal now stores values it reads as a member variable.
I modified field portal to be like this in an attempt for it
to be a more like the rest of the classes that inherit from
ArgBase. I am concerned that all classes that are returning
values for reading should return by value though.

I will have to do some tests to see if this negatively effects
scaling performance.