Tetra was broken since it was thinking workid related to input.
WorkId is actually the outputId. To get this to work properly
we presume all cells are being passed into the function, and not
a subset.

If we are passing in a subset and have a permutation, we should
use a cell array and permute it by which cells we are tetrahedralizing.