Turn off timing for basic schedule.
The serial adapter under Cygwin is running faster than the resolution of
the timer, which appears to be around the microsecond frame.  Rather than
go to great lengths trying to improve the resolution of a timer no one
cares about or make the test run longer, just disable this timer check.
There is still another timer check around a sleep command.