Remove test for counting array handle in Sine test.
The counting array handle type is fixed to dax::Id (an integer type),
which in turn fixes the output of the Sine worklet to dax::Id also.
This means that the result of sin is rounded to an integer, and I've
seen an instance where the rounding was different.  Taking the Sine of
an integer is a dumb use case anyway, so I'm deleting it.