Add a zip array handle
The zip array handle takes two array handles and provides the interface
to a single array handle with pair entries.  Of course, also had to add
our own pair class since you cannot use the std version with CUDA
devices.

The main motivation for this is to implement a key/value sort.