Add in a slice worklet and respective test to Dax.
The Slice worklet is a slight tweak on the basic marching cubes
worklet. Instead of accepting an isovalue, it instead takes an implicit
plane in the form of an origin and normal.

In the future we should look at redesigning the marching cubes and
slice worklets do use an implicit function as the input type. That
way we would have no duplication of code.