DeviceAdapterAlogirthmSerial nows uses DeviceAdapaterGeneral
Instead of implementing every single device adapter call, we
use the general class that reduces the implementation to only a couple
of algorithms. This might make the serial execution slower, but since
that is only used for debugging, nobody should care.

The upside to this is that we dramatically reduce the amount of code,
and the maintenance time for adding new methods to the device adapter.