Installed Dax 32/64 types can be overridden by project including dax.
Before this Dax at configure time would determine if it was in 32 or 64 bit mode
and configure everything at that point. This meant that if Dax was installed
on a machine, consuming application couldn't override the size of Scalar or Id.
For Dax to be a proper header only library, these things should be set by
the consuming application.

To make it easier for consuming application to set if Dax should use 32 or 64bit
data types, we now provide ConfigureFor32 and ConfigureFor64 header files.