Add a ParameterPack class to manage variable function arguments.
The new class dax::internal::ParameterPack is somewhate like boost::tuple
in that it has variable template arguments and holds an instance of each
of these types. However, this class is specifically meant to be used to
hold arguments to a function of unknown parameters. In addition to holding
the parameters, it allows you to modify the arguments (to modify the
behavior of the function) and to invoke an actual functor using the
arguments held in the class.

The main use case for this class is to simplify the implementation of the
Invoke method in the schedulars and remove the use of variable template
arguments elsewhere in Dax. It is also a convienent class to manage
signatures of real functions.