Fix UseDax*.cmake scripts
There were some minor bugs introduced when we moved from the find Boost
module to the find BoostHeaders module. The first was that UseDaxSerial
was still using the old Boost module. The second was that none of the
UseDax* scripts were loading the Dax CMake modules properly (this has
been added to DaxConfig.cmake.in to unify it).