Implemented parallel sort in generic device adapter
It uses a bitonic sort with the operations rearranged so that all
comparisons go the same way. (See the Wikipedia entry for bitonic
sorter). The sort also handles non-powers of two. Because the sort
is arranges such that all comparisons are in the same direction, any
entry outside the size but inside the next power of two is assumed
to be large.