Make mandlebulb colors a little off axis
Because the shading is for flat surfaces, when the colors are also the
same to makes disconnected surfaces look connected and certain camera
angles. This changes the coloring to be slightly off axis so that the
mapped color changes slightly and you get appropriate discontinuties.

I also increased the resolution of the color map to avoid unnecessary
jumps in the coloring.