Fix issue with error message buffers in TBB schedule.
The TBB schedule kernels catch exceptions within threads so that they
may be written to the error message buffer and later thrown in the
control environment. There was a mistake where the error message buffer
was not set in the schedule kernel (only the functor it called) so any
such exception would cause a write into a non-existing buffer.