Fix bounds of data and FPS calculation
The bounds of the data extended beyond the bounds of the Mandelbulb, so
did not produce anything interesting. Shrink the bounds to provide more
detail in the Mandelbulb surface.

The frame-per-second report was only supposed to update if the FPS
changed by more than 1, but the previous FPS being compared to was never
updated properly.